+++ repos/core/src/pages/eis/Examples.js
@@ -192,11 +192,10 @@ export class Example extends Component {
   }
 
   render () {
-    const {form, handleSubmit} = this.props
     return (
-      <ScrollView fill className='fade-in bg-neutral'>
-        <form onSubmit={handleSubmit}>
-          {this.hasData && this.hasMeta && <Render data={this.data} {...this.meta} form={form}/>}
+      <ScrollView fill className="fade-in bg-neutral">
+        <form onSubmit={this.handleSubmit}>
+          {this.hasData && this.hasMeta && <Render data={this.data} {...this.meta} form={this.form}/>}
         </form>
       </ScrollView>
     )