+++ repos/core/src/pages/eis/openl.js
@@ -1,3 +1,4 @@
+import { connect } from 'modules-pack/redux'
 import React, { Component } from 'react'
 import JsonView from 'react-ui-pack/JsonView'
 import ScrollView from 'react-ui-pack/ScrollView'
@@ -9,12 +10,17 @@ import meta from './examples/_meta'
 import { withUISetup } from './rules'
 
 // console.warn(toJSON(meta))
+const mapStateToProps = () => ({
+  initialValues: data,
+})
+
 /**
  * VIEW TEMPLATE ---------------------------------------------------------------
  * -----------------------------------------------------------------------------
  */
 // todo: adds docs about debug attribute
-@withUISetup({initialValues: data, subscription: {pristine: true, valid: true}})
+@connect(mapStateToProps)
+@withUISetup({subscription: {pristine: true, valid: true}})
 @logRender
 export default class OpenL extends Component {
   state = {
@@ -28,19 +34,18 @@ export default class OpenL extends Component {
   }
 
   render () {
-    const {form, handleSubmit} = this.props
     const {meta, showMeta} = this.state
     return (
       <>
-        <ScrollView fill className='fade-in bg-neutral'>
-          <Text className='h1 center padding bg-grey no-margin'>All Possible Configurations of The UI Render</Text>
-          <form onSubmit={handleSubmit}>
-            <Render data={this.data} {...this.meta} form={form}/>
+        <ScrollView fill className="fade-in bg-neutral">
+          <Text className="h1 center padding bg-grey no-margin">All Possible Configurations of The UI Render</Text>
+          <form onSubmit={this.handleSubmit}>
+            <Render data={this.data} {...this.meta} form={this.form}/>
           </form>
         </ScrollView>
 
         {showMeta &&
-        <ScrollView className='padding bg-neutral inverted json-tree'>
+        <ScrollView className="padding bg-neutral inverted json-tree">
           <JsonView data={meta.json} inverted/>
         </ScrollView>
         }